Adding new test case for file type check

+++ inbm/dispatcher-agent/dispatcher/aota/application_command.py
@@ -131,8 +131,8 @@ class CentOsApplication(Application):
         logger.debug(f"driver path = {driver_path}")
         try:
 
-            #if not str(driver_path).endswith('.rpm'):
-            #    raise AotaError('Invalid file type')
+            if not str(driver_path).endswith('.rpm'):
+                raise IOError('Invalid file type')
             # Remove all files in inb_driver
             for file in os.listdir(CENTOS_DRIVER_PATH):
                 remove_file(os.path.join(CENTOS_DRIVER_PATH, file))
